Add files via upload
my role was creating an interactive experience in the terminal that answers questions about the dataset

The bikeshare.py file is set up as a script that takes in raw input to create an interactive experience in the terminal that answers questions about the dataset. The experience is interactive because depending on a user's input, the answers to the questions on the previous page will change! There are four questions that will change the answers:

1-Would you like to see data for Chicago, New York, or Washington?
2-Would you like to filter the data by month, day, or not at all?
3-(If they chose month) Which month - January, February, March, April, May, or June?
4-(If they chose day) Which day - Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, or Sunday?
The answers to the questions above will determine the city and timeframe on which you'll do data analysis. After filtering the dataset, users will see the statistical result of the data, and choose to start again or exit.

I used python and I imported the pandas library to explore and analyze the data that I have in the datasets(chicago.csv, new york city.csv, washington.csv)